requirements.txt
opencv-python~=4.10.0.84
numpy~=2.0.2
pillow~=11.0.0
Flask~=3.1.0
torch~=2.5.1
torchvision~=0.20.1
pandas~=2.2.2
matplotlib~=3.9.2
kagglehub~=0.3.4
DateTime~=5.5
tqdm~=4.67.1
idx2numpy~=1.2.3
ray~=2.40.0
rembg~=2.0.0
scipy~=1.14.1
optuna~=4.1.0
diffusers~=0.31.0
accelerate~=1.2.1
onnxruntime~=1.20.1
safetensors~=0.4.5
ray[tune]~=2.40.0
ultralytics~=8.3.55